// desktop/src/ui/types.ts
export interface Config {
  ACCOUNTS_URL: string
  AI_URL?: string
  ANALYTICS_COLLECTOR_URL?: string
  BRANDING_URL?: string
  CALENDAR_URL: string
  COLLABORATOR?: string
  COLLABORATOR_URL: string
  CONFIG_URL: string
  DESKTOP_UPDATES_CHANNEL?: string // deprecated, kept for backwards compatibility of old desktop versions that will still need to be upgraded
  DESKTOP_UPDATES_CHANNELS?: string
  DESKTOP_UPDATES_URL?: string
  DISABLE_SIGNUP?: string
  HIDE_LOCAL_LOGIN?: string
  FILES_URL: string
  UPLOAD_URL: string
  DATALAKE_URL?: string
  FRONT_URL: string
  GITHUB_APP: string
  GITHUB_CLIENTID: string
  GITHUB_URL: string
  GMAIL_URL: string
  INITIAL_URL: string
  LINK_PREVIEW_URL?: string
  LIVEKIT_WS?: string
  LOVE_ENDPOINT?: string
  MODEL_VERSION: string
  PRESENCE_URL?: string
  PREVIEW_URL?: string
  PRINT_URL?: string
  PUSH_PUBLIC_KEY: string
  REKONI_URL: string
  SIGN_URL?: string
  STATS_URL?: string
  TELEGRAM_BOT_URL?: string
  TELEGRAM_URL: string
  VERSION: string
  STREAM_URL?: string
  BACKUP_URL?: string
  PUBLIC_SCHEDULE_URL?: string
  CALDAV_SERVER_URL?: string
  EXPORT_URL?: string
  MAIL_URL?: string
  COMMUNICATION_API_ENABLED?: string
  BILLING_URL?: string
  PAYMENT_URL?: string
  PULSE_URL?: string
  PASSWORD_STRICTNESS?: 'very_strict' | 'strict' | 'normal' | 'none'
  EXCLUDED_APPLICATIONS_FOR_ANONYMOUS?: string
  CLUBLAKE_URL?: string
  DISABLED_FEATURES?: string
  SIGNUP_URL?: string
}

export interface Branding {
  title?: string
  links?: {
    rel: string
    href: string
    type?: string
    sizes?: string
  }[]
  support?: {
    supportLink?: string
    reportBugLink?: string
    docsLink?: string
    privacyPolicyLink?: string
  }
  languages?: string
  lastNameFirst?: string
  defaultLanguage?: string
  defaultApplication?: string
  defaultSpace?: string
  defaultSpecial?: string
  initWorkspace?: string
  defaultInviteRole?: string
  inviteLinkGeneratorRoles?: string[]
}